+++ crates/settings_ui/src/settings_ui.rs
@@ -611,6 +611,7 @@ fn init_renderers(cx: &mut App) {
         .add_basic_renderer::<settings::ImageFileSizeUnit>(render_dropdown)
         .add_basic_renderer::<settings::StatusStyle>(render_dropdown)
         .add_basic_renderer::<settings::NotebookRunLandingMode>(render_dropdown)
+        .add_basic_renderer::<settings::NotebookFollowMode>(render_dropdown)
         .add_basic_renderer::<settings::GitPanelClickBehavior>(render_dropdown)
         .add_basic_renderer::<settings::GitPanelSortBy>(render_dropdown)
         .add_basic_renderer::<settings::GitPanelGroupBy>(render_dropdown)